rustup d528aa996 2015-04-01

+++ tests/abort.rs
@@ -1,3 +1,5 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
@@ -13,6 +15,6 @@ fn do_thing(x : u32) -> u32 {
 }
 
 fn whatever() {
-    let x : int = crust_abort();
+    let x : isize = crust_abort();
     do_thing(crust_abort());
 }
