+++ tests/autoderef1.rs
@@ -1,20 +1,22 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 use core::ops::Deref;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
 impl Deref for S {
-    type Target = uint;
-    fn deref(&self) -> &uint {
+    type Target = usize;
+    fn deref(&self) -> &usize {
         &self.x
     }
 }
 
-fn get(s: S) -> uint {
+fn get(s: S) -> usize {
     *s
 }
