+++ tests/autoderef4.rs
@@ -1,3 +1,5 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
@@ -8,15 +10,15 @@ struct S<T> {
 }
 
 struct S2 {
-    y: uint,
+    y: usize,
 }
 
 trait F {
-    fn f(&self) -> uint;
+    fn f(&self) -> usize;
 }
 
 impl F for S2 {
-    fn f(&self) -> uint {
+    fn f(&self) -> usize {
         self.y
     }
 }
@@ -28,6 +30,6 @@ impl<T> Deref for S<T> {
     }
 }
 
-fn get<T: F>(s: &S<T>) -> uint {
+fn get<T: F>(s: &S<T>) -> usize {
     s.f()
 }
