+++ tests/call1.rs
@@ -1,13 +1,15 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
-fn f(x: uint) -> uint {
+fn f(x: usize) -> usize {
     x + 1
 }
 
-fn g(x: uint) -> uint {
+fn g(x: usize) -> usize {
     f(x + 1)
 }
 
-fn crust_init() -> (uint,) { (0, ) }
+fn crust_init() -> (usize,) { (0, ) }
