+++ tests/crosscrate1.rs
@@ -1,14 +1,16 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 use core::prelude::{Option, None, Some};
 
-fn foo(x: uint) -> Option<uint> {
+fn foo(x: usize) -> Option<usize> {
     Some(x + 1)
 }
 
-fn bar(y: uint) -> Option<uint> {
+fn bar(y: usize) -> Option<usize> {
     None
 }
 
-fn crust_init() -> (uint,) { (0, ) }
+fn crust_init() -> (usize,) { (0, ) }
