+++ tests/enum2.rs
@@ -1,16 +1,18 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 mod mmm {
     enum E {
-        V1(uint),
-        V2(uint, uint),
+        V1(usize),
+        V2(usize, usize),
     }
 
-    fn mk_e(x: uint) -> E {
+    fn mk_e(x: usize) -> E {
         E::V2(x, x)
     }
 }
 
-fn crust_init() -> (uint,) { (0, ) }
+fn crust_init() -> (usize,) { (0, ) }
