+++ tests/fnfilter.rs
@@ -1,9 +1,11 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 
 extern crate core;
 
-fn fn_bad(x: int, f : |int| -> int) -> int {
+fn fn_bad(x: isize, f : |isize| -> isize) -> isize {
     f(x)
 }
 
@@ -17,8 +19,8 @@ impl<T> BadStruct<T> {
     }
 }
 
-fn fn_good(x: int) -> int {
+fn fn_good(x: isize) -> isize {
     x + 1
 }
 
-fn crust_init() -> (int,) { (0,) }
+fn crust_init() -> (isize,) { (0,) }
