+++ tests/generics1.rs
@@ -1,3 +1,5 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
@@ -6,4 +8,4 @@ fn id<T>(x: T) -> T {
     x
 }
 
-fn crust_init() -> (uint,) { (0, ) }
+fn crust_init() -> (usize,) { (0, ) }
