+++ tests/generics4.rs
@@ -1,3 +1,5 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
@@ -6,8 +8,8 @@ fn id<T>(x: T) -> T {
     x
 }
 
-fn user(x: uint) -> uint {
+fn user(x: usize) -> usize {
     id(x)
 }
 
-fn crust_init() -> (uint,) { (0, ) }
+fn crust_init() -> (usize,) { (0, ) }
