+++ tests/impl1.rs
@@ -1,15 +1,17 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
 impl S {
-    fn get(self, y: uint) -> uint {
+    fn get(self, y: usize) -> usize {
         self.x + y
     }
 }
 
-fn crust_init() -> (S, uint) { (S { x: 0 }, 0) }
+fn crust_init() -> (S, usize) { (S { x: 0 }, 0) }
