+++ tests/lifeexplicit.rs
@@ -1,13 +1,15 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
 impl S {
-    fn test<'a, 'b>(&'a self, a: &'b uint, b: &uint) -> &'a uint {
+    fn test<'a, 'b>(&'a self, a: &'b usize, b: &usize) -> &'a usize {
         &self.x
     }
 }
