+++ tests/lifeinfer.rs
@@ -1,15 +1,17 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
 impl S {
-    fn test(&self, a: &uint, b: &uint) -> &uint {
+    fn test(&self, a: &usize, b: &usize) -> &usize {
         &self.x
     }
 }
 
-fn crust_init() -> (S, uint) { (S { x: 0 }, 0) }
+fn crust_init() -> (S, usize) { (S { x: 0 }, 0) }
