+++ tests/method1.rs
@@ -1,18 +1,20 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
 impl S {
-    fn get(self) -> uint {
+    fn get(self) -> usize {
         self.x
     }
 }
 
-fn get_x(s: S) -> uint {
+fn get_x(s: S) -> usize {
     s.get()
 }
