+++ tests/method2.rs
@@ -1,3 +1,5 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
@@ -12,8 +14,8 @@ impl<T> S<T> {
     }
 }
 
-fn get_x(s: S<uint>) -> uint {
+fn get_x(s: S<usize>) -> usize {
     s.get()
 }
 
-fn crust_init() -> (S<uint>,) { (S { x: 0 },) }
+fn crust_init() -> (S<usize>,) { (S { x: 0 },) }
