+++ tests/return2.rs
@@ -1,15 +1,17 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 
 extern crate core;
 
-struct Foo { x : int }
+struct Foo { x : isize }
 
-fn foo(z: int) -> Foo {
+fn foo(z: isize) -> Foo {
     let a = Foo { x : 4i };
     Foo {
         x : return a
     }
 }
 
-fn crust_init() -> (int,) { (0i,) }
+fn crust_init() -> (isize,) { (0i,) }
