+++ tests/simple1.rs
@@ -1,10 +1,12 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
-fn foo(x: uint) -> uint {
+fn foo(x: usize) -> usize {
     let y = x + 1;
     y + 3
 }
 
-fn crust_init() -> (uint,) { (0,) }
+fn crust_init() -> (usize,) { (0,) }
