+++ tests/struct1.rs
@@ -1,12 +1,14 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
 struct S {
-    x: uint,
+    x: usize,
 }
 
-fn get_x(s: S) -> uint {
+fn get_x(s: S) -> usize {
     s.x
 }
