+++ tests/while1.rs
@@ -1,8 +1,10 @@
+#![feature(no_std)]
+#![feature(core)]
 #![crate_type = "lib"]
 #![no_std]
 extern crate core;
 
-fn f(mut x : uint) -> uint {
+fn f(mut x : usize) -> usize {
     let mut y;
     y = 4;
     while x > 0 {
